+++ src/main/java/br/com/sts/ddum/view/controllers/BuscarUserController.java
@@ -13,14 +13,22 @@ import javax.faces.event.ActionEvent;
 import javax.inject.Inject;
 
 import org.primefaces.component.tabview.TabView;
+import org.primefaces.context.RequestContext;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 
+import br.com.sts.ddum.model.entities.Repasse;
+import br.com.sts.ddum.model.entities.Responsavel;
 import br.com.sts.ddum.model.enums.ResultMessages;
 import br.com.sts.ddum.model.springsecurity.entities.Role;
 import br.com.sts.ddum.model.springsecurity.entities.User;
 import br.com.sts.ddum.service.interfaces.AclService;
+import br.com.sts.ddum.service.interfaces.RepasseService;
+import br.com.sts.ddum.service.interfaces.ResponsavelService;
 import br.com.sts.ddum.service.interfaces.UserService;
 import br.com.sts.ddum.view.utils.UtilsView;
+
+import com.google.common.collect.Sets;
 
 @Controller
 @ViewScoped
@@ -30,6 +38,12 @@ public class BuscarUserController extends BaseController {
 
 	@Inject
 	private UserService userService;
+
+	@Autowired
+	private RepasseService repasseService;
+
+	@Autowired
+	private ResponsavelService responsavelService;
 
 	@ManagedProperty("#{aclServiceImpl}")
 	private AclService aclService;
@@ -46,12 +60,14 @@ public class BuscarUserController extends BaseController {
 
 	private Role roleEdite;
 
+	private Role roleBusca;
+
 	private List<User> usuarios = new ArrayList<User>();
 
 	@PostConstruct
 	public void init() {
 		usuarios = new ArrayList<User>();
-		roleEdite = new Role();
+		roleBusca = roleEdite = new Role();
 		this.userBusca = this.userEdite = this.userRemove = new User();
 		usernameEdite = passwordEdite = new String();
 	}
@@ -60,10 +76,12 @@ public class BuscarUserController extends BaseController {
 
 		List<Role> roles = new ArrayList<Role>();
 		roles.add(roleEdite);
-		this.userEdite.setRoles(roles);
+		this.userEdite.setRoles(Sets.newHashSet(roles));
 		this.userEdite.setUsername(getUsernameEdite());
 		this.userEdite.setPassword(getPasswordEdite());
 		this.userEdite.setAtivo(true);
+		List<Responsavel> responsaveis = new ArrayList<Responsavel>();
+		this.userEdite.setResponsaveis(Sets.newHashSet(responsaveis));
 		try {
 			userService.edite(this.userEdite);
 		} catch (Exception e) {
@@ -96,6 +114,10 @@ public class BuscarUserController extends BaseController {
 			params.put("principalRole",
 					UtilsView.getControllerInstance(LoginController.class)
 							.getPrincipalRole());
+		else if (getRoleBusca() != null && getRoleBusca().getName() != null) {
+			params.put("principalRole", getRoleBusca().getName());
+		}
+
 		usuarios = userService.buscar(params);
 	}
 
@@ -107,17 +129,91 @@ public class BuscarUserController extends BaseController {
 	}
 
 	public void carregar() {
+
+		// getEditTab().setRendered(false);
+		// TabView parent = (TabView) getEditTab().getParent();
+		// parent.setActiveIndex(1);
+		LoginBean controllerInstance = UtilsView
+				.getControllerInstance(LoginBean.class);
+
+		User currentUser = controllerInstance.getCurrentUser();
+
+		if (userEdite.getId().intValue() != currentUser.getId().intValue()
+				&& !controllerInstance.getPrincipalRole().equals("ADMIN")) {
+			addErrorMessage(ResultMessages.ERROR_ONLY_ADMIN_OPERATION
+					.getDescricao());
+
+			RequestContext.getCurrentInstance().update(
+					"usuarioTabView:buscarUsuarioForm");
+			return;
+		}
+
 		getEditTab().setRendered(true);
 		TabView parent = (TabView) getEditTab().getParent();
 		int editIndex = parent.getChildren().indexOf(getEditTab());
 		parent.setActiveIndex(editIndex);
 		setUsernameEdite(userEdite.getUsername());
 		// setSenha(userEdite.getPassword());
-		setRoleEdite(userEdite.getRoles().get(0));
+		setRoleEdite(userEdite.getRoles().iterator().next());
 	}
 
 	public void remover() {
-		userService.remove(userRemove);
+
+		if (usuarioSemPermissao()) {
+
+			RequestContext.getCurrentInstance().update(
+					"usuarioTabView:buscarUsuarioForm");
+			return;
+		}
+
+		BuscarResponsavelController controllerInstance = UtilsView
+				.getControllerInstance(BuscarResponsavelController.class);
+
+		Map<String, Object> params = new HashMap<String, Object>();
+		params.put("user.id", userRemove.getId());
+		List<Responsavel> responsaveis = controllerInstance.buscar(params);
+
+		if (responsaveis == null || responsaveis.isEmpty()) {
+			try {
+				userService.remove(userRemove);
+			} catch (Exception e) {
+				if (e.getMessage() != null
+						&& e.getMessage().contains("user_username_ativo_key")) {
+					userService.forceRemove(userRemove);
+					buscar();
+					return;
+				}
+			}
+			buscar();
+			return;
+		}
+
+		Responsavel responsavel = responsaveis.get(0);
+
+		params.clear();
+		if (responsavel != null)
+			params.put("unidade.responsavel.id", responsavel.getId());
+
+		RepasseController repasseController = UtilsView
+				.getControllerInstance(RepasseController.class);
+
+		List<Repasse> repasses = repasseController.buscar(params);
+		if (repasses != null && !repasses.isEmpty()) {
+			addErrorMessage("O usuário não pode ser removido, o mesmo possui movimento(s)!");
+			return;
+		}
+
+		try {
+			userService.remove(userRemove);
+		} catch (Exception e) {
+			if (e.getMessage() != null
+					&& e.getMessage().contains("user_username_ativo_key")) {
+				userService.forceRemove(userRemove);
+				buscar();
+				return;
+			}
+		}
+		buscar();
 	}
 
 	public UserService getUserService() {
@@ -126,6 +222,22 @@ public class BuscarUserController extends BaseController {
 
 	public void setUserService(UserService userService) {
 		this.userService = userService;
+	}
+
+	public RepasseService getRepasseService() {
+		return repasseService;
+	}
+
+	public void setRepasseService(RepasseService repasseService) {
+		this.repasseService = repasseService;
+	}
+
+	public ResponsavelService getResponsavelService() {
+		return responsavelService;
+	}
+
+	public void setResponsavelService(ResponsavelService responsavelService) {
+		this.responsavelService = responsavelService;
 	}
 
 	public User getUserBusca() {
@@ -200,4 +312,11 @@ public class BuscarUserController extends BaseController {
 		this.roleEdite = roleEdite;
 	}
 
+	public Role getRoleBusca() {
+		return roleBusca;
+	}
+
+	public void setRoleBusca(Role roleBusca) {
+		this.roleBusca = roleBusca;
+	}
 }
